refactor: 플레이그라운드 쇼케이스 분리 및 페이지 헤더 네비게이션 핸들러 정리
